fix: review de la mort
