+++ app/src/lib/ui/BeneficiaryList/List.svelte
@@ -27,7 +27,6 @@
 				notebooks: [
 					{
 						id: beneficiary.notebook.id,
-						beneficiaryId: beneficiary.id,
 						members: beneficiary.notebook.members,
 					},
 				],
