+++ backend/tests/conftest.py
@@ -348,7 +348,7 @@ def get_professional_jwt() -> str:
 
 
 @pytest.fixture
-def guilia_diaby_jwt() -> str:
+def giulia_diaby_jwt() -> str:
     return "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJodHRwczovL2hhc3VyYS5pby9qd3QvY2xhaW1zIjp7IngtaGFzdXJhLWFsbG93ZWQtcm9sZXMiOlsib3JpZW50YXRpb25fbWFuYWdlciJdLCJ4LWhhc3VyYS1kZWZhdWx0LXJvbGUiOiJvcmllbnRhdGlvbl9tYW5hZ2VyIiwieC1oYXN1cmEtdXNlci1pZCI6IjJhZGRkMTBmLTliZDMtNGQzNy1iM2M5LTEwYTZlMmM0YmU0ZiIsIngtaGFzdXJhLW9yaWVudGF0aW9uTWFuYWdlci1pZCI6IjYwN2NiNmY4LTllMzMtNGNlOC05OGIxLTM4ZTYwYzlkZGE5OSIsIngtaGFzdXJhLWRlcGxveW1lbnQtaWQiOiI0ZGFiODAzNi1hODZlLTRkNWYtOWJkNC02Y2U4OGMxOTQwZDAifSwiaWQiOiIyYWRkZDEwZi05YmQzLTRkMzctYjNjOS0xMGE2ZTJjNGJlNGYiLCJyb2xlIjoib3JpZW50YXRpb25fbWFuYWdlciIsImRlcGxveW1lbnRJZCI6IjRkYWI4MDM2LWE4NmUtNGQ1Zi05YmQ0LTZjZTg4YzE5NDBkMCIsIm9yaWVudGF0aW9uTWFuYWdlcklkIjoiNjA3Y2I2ZjgtOWUzMy00Y2U4LTk4YjEtMzhlNjBjOWRkYTk5IiwiaWF0IjoxNjY5MzAyNzYxLCJleHAiOjE5ODQ4Nzg3NjEsInN1YiI6IjJhZGRkMTBmLTliZDMtNGQzNy1iM2M5LTEwYTZlMmM0YmU0ZiJ9.X-4PFqGMLszH2_x1wSUkV-cME0Ln1WljeSMnyQ1LJA0"
