+++ hasura/metadata/databases/carnet_de_bord/tables/public_orientation_request.yaml
@@ -102,29 +102,6 @@ insert_permissions:
         - requestor_account_id
         - status
 select_permissions:
-  - role: admin_structure
-    permission:
-      columns:
-        - decided_at
-        - id
-        - status
-      filter:
-        beneficiary:
-          structures:
-            structure:
-              admins:
-                admin_structure_id:
-                  _eq: X-Hasura-AdminStructure-Id
-  - role: manager
-    permission:
-      columns:
-        - decided_at
-        - id
-        - status
-      filter:
-        beneficiary:
-          deployment_id:
-            _eq: X-Hasura-Deployment-Id
   - role: orientation_manager
     permission:
       columns:
