+++ scripts/check-app-codegen.sh
@@ -5,19 +5,18 @@ function get_file_modification_timestamp() {
 	local pattern=$1
 	local fs_ts=$(for f in $(git status --porcelain "$pattern"|cut -c4-); do
 		if [[ -f "$f" ]]; then date -r "$f" '+%s'; else date '+%s'; fi # for deleted files, we use the current date
-	done)
-	local commit_ts=$(git log -1 --format=%ct $pattern)
+	done | sort -rg | head -1)
+	local commit_ts=$(git log -1 --format=%ct "$pattern")
   if [[ "$commit_ts" -gt "$fs_ts" ]]; then
 		echo "$commit_ts"
 	else
 		echo "$fs_ts"
 	fi
 }
-patterns=('hasura/metadata/**/*.yaml' 'hasura/metadata/**/*.graphql' 'app/**/*.gql')
-for pattern in $patterns; do
-	gql_ts="$(get_file_modification_timestamp $pattern)"
-	generated_file_ts="$(get_file_modification_timestamp 'app/src/lib/graphql/_gen/typed-document-nodes.ts')"
 
+for pattern in 'app/*.gql'; do
+	gql_ts="$(get_file_modification_timestamp "$pattern")"
+	generated_file_ts="$(get_file_modification_timestamp 'app/src/lib/graphql/_gen/typed-document-nodes.ts')"
 	if [[ "$gql_ts" -gt "$generated_file_ts" ]]; then
 		echo "Generated file is older than source file: codegen needs to be ran"
 		exit 1
